Renamed DuotoneImage.tsx to duotone-image.tsx and removed commented code from about-hero.tsx
