+++ src/components/duotone-image.tsx
